pyproject.toml
[project]
name = "myagent"
version = "0.1.0"
description = "AgentScope Skill/Tool/MCP Validation Platform"
requires-python = ">=3.11"
dependencies = [
    "agentscope-runtime==1.1.3",
    "httpx>=0.28.1",
    "pydantic-settings>=2.0",
    "python-frontmatter>=1.1.0",
]

[dependency-groups]
dev = [
    "fakeredis>=2.31.0",
    "httpx==0.28.1",
    "pytest==9.0.3",
]

[tool.pytest.ini_options]
pythonpath = ["."]
